Move Landing to components root
