+++ components/index.jsx
@@ -1,2 +1,1 @@
-export { default as Landing } from './Landing'
 export { default as WalletView } from './Wallet'
